requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2
django-allauth==65.7.0
django-environ==0.12.0
django-htmx==1.23.0
django-widget-tweaks==1.5.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
ruff==0.11.4
sqlparse==0.5.3
tzdata==2025.2
